in order to clena the scripts
